Move const eval query components into their own module
