+++ src/librustc_mir/const_eval/error.rs
@@ -1,8 +1,8 @@
 use std::error::Error;
 use std::fmt;
 
-use crate::interpret::InterpErrorInfo;
-
+use super::InterpCx;
+use crate::interpret::{ConstEvalErr, InterpErrorInfo, Machine};
 #[derive(Clone, Debug)]
 pub enum ConstEvalError {
     NeedsRfc(String),
@@ -28,3 +28,15 @@ impl fmt::Display for ConstEvalError {
 }
 
 impl Error for ConstEvalError {}
+
+/// Turn an interpreter error into something to report to the user.
+/// As a side-effect, if RUSTC_CTFE_BACKTRACE is set, this prints the backtrace.
+/// Should be called only if the error is actually going to to be reported!
+pub fn error_to_const_error<'mir, 'tcx, M: Machine<'mir, 'tcx>>(
+    ecx: &InterpCx<'mir, 'tcx, M>,
+    mut error: InterpErrorInfo<'tcx>,
+) -> ConstEvalErr<'tcx> {
+    error.print_backtrace();
+    let stacktrace = ecx.generate_stacktrace(None);
+    ConstEvalErr { error: error.kind, stacktrace, span: ecx.tcx.span }
+}
